Updated driver factory and some refactoring
