--- PageObjects/Framework/WebDriver/WebDriverTypes.cs
@@ -1,8 +1,0 @@
-﻿namespace PageObjects.Framework.WebDriver
-{
-    public enum WebDriverTypes
-    {
-        Chrome,
-        FireFox
-    }
-}
